update - criada funcionalidade de pergunta se deseja excluir cliente, ajuste na função app/client/views.py client_delete e criado em templates client_delete.html

+++ apps/clients/views.py
@@ -43,10 +43,12 @@ def client_edit(request, id_client):
 def client_delete(request, id_client):
     if not request.user.is_authenticated:
             return redirect('login')
-
-    id_client = models.Client_Register.objects.get(id=id_client)
-    id_client.delete()
-    messages.success(request, 'Deleção realizada com sucesso!')
-
-    return redirect('client_list')
     
+    if request.method == 'POST':
+        id_client = models.Client_Register.objects.get(id=id_client)
+        id_client.delete()
+        messages.success(request, 'Deleção realizada com sucesso!')
+        return redirect('client_list')
+    
+    return render(request, 'clients/client_delete.html', {'id_client': id_client})
+
